C4-71 move split /compound_search from /search
